pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aser"
version = "0.1.6"
description = "An AI Agent"
authors = [
    {name = "rickey", email = "hello_rickey@163.com"},
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13.2"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "openai >= 1.63.2", 
    "web3 >= 7.8.0",
    "discord-py-interactions >= 5.14.0",
    "python-telegram-bot >= 21.10",
    "tweepy >= 4.15.0",
    "supabase >= 2.13.0",
    "tinydb >= 4.8.2",
    "fastapi >= 0.115.8",
    "uvicorn >= 0.34.0",
    "python-dotenv >= 1.0.1",
    "chromadb >= 0.6.3",
    "tiktoken >= 0.9.0",
    "canonicaljson >= 2.0.0",
    "mastodon.py >= 2.0.1",
    "atproto >=0.0.59",
    "rich >= 13.9.4",
    "bs4 >=0.0.2",
    "html2text >= 2025.4.15",
    "semchunk >= 3.2.3", 
    "fastmcp >= 2.12.0"
]

[project.urls]
Homepage = "https://github.com/AmeNetwork/aser"
"Bug Tracker" = "https://github.com/AmeNetwork/aser/issues"

[tool.setuptools]
packages = ["aser"]
include-package-data = true
license-files = []

[tool.setuptools.package-data]
"aser" = ["**/*.py", "**/*.json", "**/*.yaml", "**/*.yml"]

